+++ packages/components/index.d.ts
@@ -390,6 +390,7 @@ interface InputBasicProps {
   color?: Color;
   textColor?: Color;
   colorFocus?: Color;
+  colorError?: Color;
   size?: 'medium' | 'large';
   mobileSize?: 'medium' | 'large';
   inputProps?: React.InputHTMLAttributes<HTMLInputElement>;