+++ packages/components/src/components/text_field/input.jsx
@@ -86,9 +86,13 @@ export default class Input extends PureComponent {
      */
     textColor: PropTypes.string,
     /**
-     * Component focus color from theme
+     * Component focus color from the theme.
      */
     colorFocus: PropTypes.string,
+    /**
+     * Component error color from the theme.
+     */
+    colorError: PropTypes.string,
     /**
      * Component size.
      */
@@ -131,13 +135,13 @@ export default class Input extends PureComponent {
     valid: true,
     placeholder: undefined,
     placeholderColor: 'grey_4',
-    // label: undefined,
     children: null,
     name: undefined,
     bgType: 'fill',
     color: 'light_grey',
     textColor: 'black',
     colorFocus: 'primary',
+    colorError: 'wrong',
     size: 'medium',
     mobileSize: undefined,
     inputProps: {},
@@ -207,13 +211,13 @@ export default class Input extends PureComponent {
       valid,
       placeholder,
       placeholderColor,
-      // label,
       children,
       name,
       bgType,
       color,
       textColor,
       colorFocus,
+      colorError,
       size: propsSize,
       mobileSize,
       inputProps,
@@ -227,16 +231,6 @@ export default class Input extends PureComponent {
       ...otherInputProps
     } = inputProps;
     const Component = multiLine ? 'textarea' : 'input';
-
-    // {label && (
-    //   <div
-    //     onClick={this.setFocus}
-    //     className={s.label}
-    //     ref={(node) => { this.labelNode = node; }}
-    //   >
-    //     {label}
-    //   </div>
-    // )}
 
     let size = propsSize || Input.defaultProps.size;
 
@@ -265,13 +259,15 @@ export default class Input extends PureComponent {
             'input_field',
             'round_small',
             [`input_field_${size}`],
-            [`stroke_${color}`],
             {
+              [`stroke_${color}`]: valid,
+              [`stroke_${colorError}`]: !valid,
+              [`input_field_focus_${colorFocus}`]: valid,
+
               fill_white: bgType === 'stroke',
               [`fill_${color}`]: bgType === 'fill',
               [`text_${textColor}`]: textColor,
             },
-            [`input_field_focus_${colorFocus}`],
             [`input_placeholder_color_${placeholderColor}`],
             classNameInput,
           )}