+++ packages/docs-app/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.14](https://github.com/PeculiarVentures/peculiar-react-components/compare/docs-app@1.0.13...docs-app@1.0.14) (2020-12-16)
+
+**Note:** Version bump only for package docs-app
+
+
+
+
+
+## [1.0.13](https://github.com/PeculiarVentures/peculiar-react-components/compare/docs-app@1.0.12...docs-app@1.0.13) (2020-12-16)
+
+**Note:** Version bump only for package docs-app
+
+
+
+
+
+## [1.0.12](https://github.com/PeculiarVentures/peculiar-react-components/compare/docs-app@1.0.11...docs-app@1.0.12) (2020-11-29)
+
+**Note:** Version bump only for package docs-app
+
+
+
+
+
+## [1.0.11](https://github.com/PeculiarVentures/peculiar-react-components/compare/docs-app@1.0.10...docs-app@1.0.11) (2020-11-26)
+
+**Note:** Version bump only for package docs-app
+
+
+
+
+
 ## [1.0.10](https://github.com/PeculiarVentures/peculiar-react-components/compare/docs-app@1.0.9...docs-app@1.0.10) (2020-11-12)
 
 **Note:** Version bump only for package docs-app